--- posthog-node/index.ts
@@ -1,4 +1,0 @@
-export * from './src/extensions/sentry-integration'
-export * from './src/extensions/express'
-// Uses node as the default runtime
-export * from 'posthog-node/runtimes/index.node'
