+++ posthog-node/package.json
@@ -19,8 +19,8 @@
     "email": "hey@posthog.com",
     "url": "https://posthog.com"
   },
-  "main": "lib/index.cjs.js",
-  "module": "lib/index.esm.js",
+  "main": "lib/node/index.cjs.js",
+  "module": "lib/node/index.esm.js",
   "types": "lib/index.d.ts",
   "dependencies": {
     "axios": "^1.8.2"
@@ -44,6 +44,11 @@
         "require": "./lib/edge/index.cjs.js",
         "default": "./lib/edge/index.js"
       },
+      "node": {
+        "import": "./lib/node/index.esm.js",
+        "require": "./lib/node/index.cjs.js",
+        "default": "./lib/node/index.js"
+      },
       "edge-light": {
         "import": "./lib/edge/index.esm.js",
         "require": "./lib/edge/index.cjs.js",
@@ -54,8 +59,8 @@
         "require": "./lib/edge/index.cjs.js",
         "default": "./lib/edge/index.js"
       },
-      "import": "./lib/index.esm.js",
-      "require": "./lib/index.cjs.js"
+      "import": "./lib/node/index.esm.js",
+      "require": "./lib/node/index.cjs.js"
     }
   }
 }
