--- posthog-node/runtimes/index.edge.ts
@@ -1,13 +1,0 @@
-import { defaultStackParser } from '../src/extensions/error-tracking/stack-parser.edge'
-import ErrorTracking from '../src/error-tracking'
-
-import { PostHogBackendClient } from '../src/client'
-
-ErrorTracking.stackParser = defaultStackParser
-ErrorTracking.frameModifiers = []
-
-export class PostHog extends PostHogBackendClient {
-  getLibraryId(): string {
-    return 'posthog-edge'
-  }
-}
