+++ posthog-node/src/entrypoints/index.edge.ts
@@ -1,0 +1,15 @@
+export * from '../exports'
+
+import { defaultStackParser } from '../extensions/error-tracking/stack-parser.edge'
+import ErrorTracking from '../extensions/error-tracking'
+
+import { PostHogBackendClient } from '../client'
+
+ErrorTracking.stackParser = defaultStackParser
+ErrorTracking.frameModifiers = []
+
+export class PostHog extends PostHogBackendClient {
+  getLibraryId(): string {
+    return 'posthog-edge'
+  }
+}
