+++ posthog-node/src/exports.ts
@@ -1,0 +1,3 @@
+export * from './extensions/sentry-integration'
+export * from './extensions/express'
+export * from './types'
