+++ posthog-node/src/extensions/error-tracking/index.ts
@@ -1,9 +1,9 @@
-import { EventHint, StackFrameModifierFn, StackParser } from './extensions/error-tracking/types'
-import { addUncaughtExceptionListener, addUnhandledRejectionListener } from './extensions/error-tracking/autocapture'
-import { PostHogBackendClient, PostHogOptions } from './client'
+import { EventHint, StackFrameModifierFn, StackParser } from './types'
+import { addUncaughtExceptionListener, addUnhandledRejectionListener } from './autocapture'
+import { PostHogBackendClient } from '../../client'
 import { uuidv7 } from 'posthog-core/src/vendor/uuidv7'
-import { propertiesFromUnknownInput } from './extensions/error-tracking/error-conversion'
-import { EventMessage } from './types'
+import { propertiesFromUnknownInput } from './error-conversion'
+import { EventMessage, PostHogOptions } from '../../types'
 
 const SHUTDOWN_TIMEOUT = 2000
