+++ posthog-node/src/extensions/feature-flags/crypto-helpers.ts
@@ -1,5 +1,5 @@
 /// <reference lib="dom" />
-import { Lazy } from 'posthog-node/src/lazy'
+import { Lazy } from 'posthog-node/src/extensions/feature-flags/lazy'
 
 const nodeCrypto = new Lazy(async () => {
   try {
