+++ posthog-node/tsconfig.json
@@ -1,6 +1,8 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
+    "incremental": false,
+    "declaration": false,
     "types": ["node"],
     "typeRoots": ["./node_modules/@types", "../node_modules/@types"]
   }
